wip: conjunction-ranked: try to optimize tests
No improvement: we are still slower on vcsn score.

 2.35s : a.conjunction(a)     # a = std([a-e]?{50}), c = [a-e] -> Z
 2.64s : a.conjunction2(a)    # a = std([a-e]?{50}), c = [a-e] -> Z
 2.36s : a.conjunction(a)     # a = std([a-e]?{50}), c = [a-e]? -> Z
 2.67s : a.conjunction2(a)    # a = std([a-e]?{50}), c = [a-e]? -> Z
 0.78s : a.conjunction(a)     # a = thompson([a-e]?{80})
 0.78s : a.conjunction2(a)    # a = thompson([a-e]?{80})
 1.53s : a.conjunction(a, a, a) # a = thompson([a-e]?{5})
 1.62s : a.conjunction2(a, a, a) # a = thompson([a-e]?{5})

However, on random automata with spontaneous transitions, we are much better.

       before after %
         us    us
0 EQUAL  93    55    59
1 EQUAL  92    65    71
2 EQUAL  77    56    72
3 EQUAL  91    60    66
4 EQUAL  73    47    65
5 EQUAL  85    55    65
6 EQUAL 116    82    71
7 EQUAL  71    41    58
8 EQUAL  94    66    70
9 EQUAL  64    44    69
10 EQUAL 83    54    64
11 EQUAL 80    53    67
12 EQUAL 81    55    68

* libexec/vcsn-score: .
* vcsn/algos/conjunction-ranked.hh: .
* vcsn/core/tuple-automaton.hh: .

+++ vcsn/core/tuple-automaton.hh
@@ -44,7 +44,7 @@ namespace vcsn
       /// State names: Tuple of states of input automata.
       using state_name_t
         = std::conditional_t<Ranked,
-                             std::tuple<state_t_of<Auts>..., unsigned>,
+                             std::tuple<state_t_of<Auts>..., unsigned short>,
                              std::tuple<state_t_of<Auts>...>>;
 
       /// The type of context of the result.
@@ -239,7 +239,7 @@ namespace vcsn
       }
 
       template <bool Lazy = false, bool Ranked_ = Ranked>
-      auto state(state_t_of<Auts>... ss, unsigned rank)
+      auto state(state_t_of<Auts>... ss, unsigned short rank)
         -> std::enable_if_t<Ranked_, state_t>
       {
         return state<Lazy>(std::make_tuple(ss..., rank));
